Refactor/freight model (#423)
* First draft of new freight_purpose module

* Change name to commodity

* Fix imports
